+++ C11/ex05/do_op.h
@@ -6,7 +6,7 @@
 /*   By: yaharkat <yaharkat@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/03 17:35:23 by yaharkat          #+#    #+#             */
-/*   Updated: 2023/09/03 18:28:02 by yaharkat         ###   ########.fr       */
+/*   Updated: 2023/09/04 20:17:56 by yaharkat         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,14 +22,14 @@ typedef enum s_bool
 }		t_bool;
 
 void	ft_putnbr(long nb);
-int		ft_atoi(char *str);
-void	do_op(int a, int b, char op);
-int		ft_strcmp(char *s1, char *s2);
+long	ft_atoi(char *str);
+void	do_op(long a, long b, char op);
+long	ft_strcmp(char *s1, char *s2);
 void	ft_putchar(char c);
-int		ft_multiplication(int a, int b);
-int		ft_addition(int a, int b);
-int		ft_substraction(int a, int b);
-int		ft_division(int a, int b);
-int		ft_modulo(int a, int b);
+long	ft_multiplication(long a, long b);
+long	ft_addition(long a, long b);
+long	ft_substraction(long a, long b);
+long	ft_division(long a, long b);
+long	ft_modulo(long a, long b);
 
 #endif
