+++ C11/ex05/op.c
@@ -6,31 +6,31 @@
 /*   By: yaharkat <yaharkat@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/03 16:35:35 by yaharkat          #+#    #+#             */
-/*   Updated: 2023/09/03 17:31:36 by yaharkat         ###   ########.fr       */
+/*   Updated: 2023/09/04 20:16:52 by yaharkat         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_multiplication(int a, int b)
+long	ft_multiplication(long a, long b)
 {
 	return (a * b);
 }
 
-int	ft_addition(int a, int b)
+long	ft_addition(long a, long b)
 {
 	return (a + b);
 }
 
-int	ft_substraction(int a, int b)
+long	ft_substraction(long a, long b)
 {
 	return (a - b);
 }
 
-int	ft_division(int a, int b)
+long	ft_division(long a, long b)
 {
 	return (a / b);
 }
 
-int	ft_modulo(int a, int b)
+long	ft_modulo(long a, long b)
 {
 	return (a % b);
 }
